feat: Add randomInt function to number.utils.ts
This commit adds a new function called randomInt to the number.utils.ts file. The randomInt function generates a random integer between a given minimum and maximum value. This function can be used for various purposes, such as generating random numbers for testing or generating random indexes for array manipulation.

Co-authored-by: [Author Name]